creating donors folder and moving files there
